Java: use cmake plugin to build FFI sources.
