+++ java/core/.gitignore
@@ -1,0 +1,3 @@
+build/
+cmake-build/
+
