+++ java/core/src/main/cpp/ffi/ffi.cc
@@ -19,15 +19,19 @@ limitations under the License.
 #include <string.h>
 #include <sys/mman.h>
 
-#include "fling.h"
+#include "vineyard/common/memory/fling.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /*
  * Class:     io_v6d_core_common_memory_ffi_Fling
  * Method:    sendFD
  * Signature: (II)I
  */
-JNIEXPORT jint JNICALL Java_io_v6d_core_common_memory_ffi_Fling_sendFD
-  (JNIEnv *, jclass, jint conn, jint fd) {
+JNIEXPORT jint JNICALL Java_io_v6d_core_common_memory_ffi_Fling_sendFD(
+    JNIEnv*, jclass, jint conn, jint fd) {
   return send_fd(conn, fd);
 }
 
@@ -36,8 +40,8 @@ JNIEXPORT jint JNICALL Java_io_v6d_core_common_memory_ffi_Fling_sendFD
  * Method:    recvFD
  * Signature: (I)I
  */
-JNIEXPORT jint JNICALL Java_io_v6d_core_common_memory_ffi_Fling_recvFD
-  (JNIEnv *, jclass, jint conn) {
+JNIEXPORT jint JNICALL
+Java_io_v6d_core_common_memory_ffi_Fling_recvFD(JNIEnv*, jclass, jint conn) {
   return recv_fd(conn);
 }
 
@@ -46,15 +50,20 @@ JNIEXPORT jint JNICALL Java_io_v6d_core_common_memory_ffi_Fling_recvFD
  * Method:    mapSharedMem
  * Signature: (IJZZ)J
  */
-JNIEXPORT jlong JNICALL Java_io_v6d_core_common_memory_ffi_Fling_mapSharedMem
-  (JNIEnv *, jclass, jint fd, jlong map_size, jboolean readonly, jboolean realign) {
+JNIEXPORT jlong JNICALL Java_io_v6d_core_common_memory_ffi_Fling_mapSharedMem(
+    JNIEnv*, jclass, jint fd, jlong map_size, jboolean readonly,
+    jboolean realign) {
   size_t length = map_size;
   if (realign == JNI_TRUE) {
     length -= sizeof(size_t);
   }
-  void *pointer = mmap(NULL, length, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+  void* pointer = mmap(NULL, length, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
   if (pointer == MAP_FAILED) {
     fprintf(stderr, "mmap failed: errno = %d: %s\n", errno, strerror(errno));
   }
   return reinterpret_cast<jlong>(pointer);
 }
+
+#ifdef __cplusplus
+}
+#endif
